+++ packages/playground/content/section.mdx
