+++ packages/playground/content/show-references.mdx
