+++ packages/playground/content/simple-code.mdx
