+++ packages/playground/content/simple-editor.mdx
