+++ packages/playground/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@code-hike/mdx-test",
+  "name": "@code-hike/playground",
   "version": "0.3.0-next.0",
   "private": "true",
   "scripts": {
